Split Controller tests into Card and Note Test classes
